@@ -1,9 +1,9 @@
 {
     "name": "kevinpapst/lockdown-per-user-plugin",
-    "description": "A Kimai plugin that allows to set the lockdown period per user.",
-    "homepage": "https://www.kimai.org/store/translation-bundle.html",
+    "description": "Allows to set the lockdown period per user.",
+    "homepage": "https://www.kevinpapst.de",
     "type": "kimai-plugin",
-    "version": "1.0",
+    "version": "2.0",
     "keywords": [
         "kimai",
         "kimai-plugin"
@@ -12,14 +12,13 @@
     "authors": [
         {
             "name": "Kevin Papst",
-            "email": "kevin@kevinpapst.de",
             "homepage": "https://www.kevinpapst.de"
         }
     ],
     "extra": {
         "kimai": {
-            "require": "1.16.9",
-            "name": "LockdownPerUserBundle"
+            "require": 20000,
+            "name": "Lockdown per user"
         }
     },
     "autoload": {
@@ -30,10 +29,11 @@
     "config": {
         "allow-plugins": {
             "composer/package-versions-deprecated": false,
-            "symfony/flex": false
+            "symfony/flex": false,
+            "symfony/runtime": false
         },
         "platform": {
-            "php": "7.3"
+            "php": "8.1.3"
         },
         "preferred-install": {
             "*": "dist"
@@ -51,18 +51,13 @@
             "@phpstan"
         ]
     },
-    "require": {
-        "doctrine/dbal": "^2.1"
-    },
     "require-dev": {
-        "doctrine/doctrine-fixtures-bundle": "^3.3",
-        "fakerphp/faker": "^1.15",
         "friendsofphp/php-cs-fixer": "^3.0",
-        "kevinpapst/kimai2": "^1.15",
+        "kimai/kimai": "2.0.x-dev",
         "phpstan/phpstan": "^1.0",
         "phpstan/phpstan-doctrine": "^1.0",
         "phpstan/phpstan-symfony": "^1.0",
-        "symfony/console": "^4.0",
-        "symfony/event-dispatcher": "^4.0"
+        "symfony/console": "^6.0",
+        "symfony/event-dispatcher": "^6.0"
     }
 }
